Decided to use the same Composable used on OverviewScreen, ListInfo. Added extra parameter to ListInfo that hides/displays option Button. Created Composable to display ItemLists in AddScreen, but nothing is hooked up.
