+++ app/src/main/res/values/dimens.xml
@@ -6,6 +6,7 @@
     <!-- AddScreen (as) -->
     <dimen name="as_padding_all">16.dp</dimen>
     <dimen name="as_query_padding_bottom">12.dp</dimen>
+    <dimen name="as_list_spacedBy">6.dp</dimen>
     <!-- AddScreen.BottomSheet (asbs) -->
     <dimen name="asbs_vertical_spacedBy">4.dp</dimen>
     <dimen name="asbs_button_padding_horizontal">16.dp</dimen>
